+++ ravyn/core/terminal/utils.py
@@ -5,6 +5,8 @@ from rich_toolkit import RichToolkit, RichToolkitTheme
 from rich_toolkit.styles import TaggedStyle
 from uvicorn.logging import DefaultFormatter
 
+from ravyn.conf import settings
+
 
 class CustomFormatter(DefaultFormatter):
     def __init__(self, *args: Any, **kwargs: Any) -> None:
@@ -54,9 +56,6 @@ def get_log_config() -> dict[str, Any]:
     }
 
 
-logger = logging.getLogger(__name__)
-
-
 def get_ui_toolkit() -> RichToolkit:
     """
     Returns a RichToolkit instance with a custom theme for the terminal UI.
@@ -74,17 +73,7 @@ def get_ui_toolkit() -> RichToolkit:
     """
     theme = RichToolkitTheme(
         style=TaggedStyle(tag_width=11),
-        theme={
-            "tag.title": "white on #f11e64",
-            "tag": "white on #af1b4c",
-            "placeholder": "grey85",
-            "text": "white",
-            "selected": "#e92063",
-            "result": "grey85",
-            "progress": "on #af1b4c",
-            "error": "red",
-            "log.info": "black on blue",
-        },
+        theme=settings.runserver_theme,
     )
 
     return RichToolkit(theme=theme)